Update new design for both login & register pages
- update router
- update login.scss
- discard old design
