+++ client/src/assets/styles/login/login.scss
@@ -1,45 +1,178 @@
 .entrance {
 	overflow: hidden;
 	text-align: center;
-	// padding-top: 100px;
 
-	:deep(#chat-logo) {
-		width: 300px;
-		height: 252px;
-		margin-bottom: 20px;
-		margin-left: 35px;
+	*,
+	::after,
+	::before {
+		box-sizing: border-box;
+		font-weight: 300;
 	}
 
-	.image-slot {
-		width: 300px;
-		height: 252px;
-		margin-bottom: 20px;
-		margin-left: 35px;
+	// $bg_body: #212121;
+	$bg_body: $themeColorDeep;
+	$bg_gradient: linear-gradient(
+		90deg,
+		$emphasisColorA 0%,
+		$emphasisColorA 37%,
+		$fontColorDeep 94%
+	);
+
+	* {
+		// background-color: $bg_body;
+		color: #fff;
+		// font-family: monospace, serif;
+		letter-spacing: 0.05em;
 	}
 
-	.signin,
-	.register {
-		overflow: hidden;
-		text-align: center;
-		padding: 0px 30% 0px;
+	h1 {
+		font-size: 23px;
+	}
 
-		.sign-in-link {
-			display: inline-block;
-			text-decoration: underline;
-			color: $fontColorDeep;
-			font-size: 12px;
+	.login, .signup {
+		width: 300px;
+		padding: 64px 15px 24px;
+		margin: 0 auto;
+		.control {
+			margin: 0 0 24px;
+			input {
+				width: 100%;
+				padding: 14px 16px;
+				border: 0;
+				background: transparent;
+				color: #fff;
+				// font-family: monospace, serif;
+				letter-spacing: 0.05em;
+				font-size: 16px;
+				&:hover,
+				&:focus {
+					outline: none;
+					border: 0;
+				}
+			}
 		}
-
-		.el-button {
-			width: 250px;
-			border-radius: 25px;
-			border: 0px;
-			color: $fontColorDeep;
-			background-color: $themeColorLight;
+		// 输入框格式错误显示红色点线
+		.error-prompt {
+			text-decoration: underline dashed rgb(255, 0, 89) 2.5px;
 		}
+		.btn {
+			// margin: 0 0 24px;
+			width: 100%;
+			display: block;
+			padding: 14px 16px;
+			background: transparent;
+			outline: none;
+			border: 0;
+			color: #fff;
+			letter-spacing: 0.1em;
+			font-weight: bold;
+			font-size: 16px;
+		}
+		.or-sign-in {
+			text-decoration: underline $fontColorLight 1.5px;
+			text-align: center;
+			&:hover {
+				cursor: pointer;
+			}
+		}
+	}
 
-		:deep(.el-form .el-form-item label) {
-			color: $fontColorDeep;
+	.block-cube {
+		position: relative;
+		.bg-top {
+			position: absolute;
+			height: 10px;
+			background: rgb(2, 0, 36);
+			background: $bg_gradient;
+			bottom: 100%;
+			left: 5px;
+			right: -5px;
+			transform: skew(-45deg, 0);
+			margin: 0;
+			.bg-inner {
+				bottom: 0;
+			}
+		}
+		.bg {
+			position: absolute;
+			left: 0;
+			top: 0;
+			right: 0;
+			bottom: 0;
+			background: rgb(2, 0, 36);
+			background: $bg_gradient;
+		}
+		.bg-right {
+			position: absolute;
+			background: rgb(2, 0, 36);
+			background: $fontColorDeep;
+			top: -5px;
+			z-index: 0;
+			bottom: 5px;
+			width: 10px;
+			left: 100%;
+			transform: skew(0, -45deg);
+			.bg-inner {
+				left: 0;
+			}
+		}
+		.bg {
+			.bg-inner {
+				transition: all 0.2s ease-in-out;
+			}
+		}
+		.bg-inner {
+			background: $bg_body;
+			position: absolute;
+			left: 2px;
+			top: 2px;
+			right: 2px;
+			bottom: 2px;
+		}
+		.text {
+			position: relative;
+			z-index: 2;
+		}
+		&.block-input {
+			input {
+				position: relative;
+				z-index: 2;
+				&:focus ~ .bg-right .bg-inner,
+				&:focus ~ .bg-top .bg-inner,
+				&:focus ~ .bg-inner .bg-inner {
+					top: 100%;
+					background: rgba(255, 255, 255, 0.5);
+				}
+			}
+			.bg-top,
+			.bg-right,
+			.bg {
+				background: rgba(255, 255, 255, 0.5);
+				transition: background 0.2s ease-in-out;
+			}
+			.bg-right,
+			.bg-top {
+				.bg-inner {
+					transition: all 0.2s ease-in-out;
+				}
+			}
+			&:focus,
+			&:hover {
+				.bg-top,
+				.bg-right,
+				.bg {
+					background: rgba(255, 255, 255, 0.8);
+				}
+			}
+		}
+		// State hover, focus
+		&.block-cube-hover:focus,
+		&.block-cube-hover:hover {
+			.bg {
+				.bg-inner {
+					top: 100%;
+				}
+			}
 		}
 	}
 }
