+++ client/src/views/login/index.vue
@@ -1,11 +1,5 @@
 <template>
 	<div class="entrance">
-		<div id="chat-logo"></div>
-		<!-- <el-image id="chat-logo" :src="imgUrl">
-			<template #placeholder>
-				<div class="image-slot">加载中<span class="dot">...</span></div>
-			</template>
-		</el-image> -->
 		<router-view />
 	</div>
 </template>
@@ -13,9 +7,7 @@
 <script>
 	export default {
 		data() {
-			return {
-				imgUrl: require('@/assets/chat-logo-trans.png'),
-			};
+			return {};
 		},
 		mounted() {},
 	};
